+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,8 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-top-sites**
+  * Added `isDefault` to the top site entity, which allows application to specify a default top site that is added by the application. This is called through `TopSiteStorage.addTopSite`.
 
 # 39.0.0
 
@@ -54,6 +56,12 @@ permalink: /changelog/
 
 * **feature-media**
   * Muted media will not start the media service anymore, causing no media notification to be shown and no audio focus getting requested.
+
+* **feature-fullscreen**
+  * ⚠️ **This is a breaking change**: Added `viewportFitChanged` to support Android display cutouts.
+
+  * **feature-qr**
+    * Moved `AutoFitTextureView` from `support-base` to `feature-qr`.
 
 * **feature-session**
   * ⚠️ **This is a breaking change**: Added `viewportFitChanged` to `FullScreenFeature` for supporting Android display cutouts.